@@ -14,43 +14,44 @@
         "ezsystems/ezpublish-community": "*"
     },
     "autoload": {
-        "psr-0": { "": "src/" }
+      "psr-4": { "": "src/" },
+      "classmap": [ "app/AppKernel.php", "app/AppCache.php" ]
+    },
+    "autoload-dev": {
+        "psr-4": { "Tests\\": "tests/" }
     },
     "require": {
         "php": "~5.6|~7.0",
-        "symfony/symfony": "~2.8.0",
-        "twig/extensions": "~1.0",
-        "symfony/assetic-bundle": "~2.3",
-        "symfony/swiftmailer-bundle": "~2.3",
-        "symfony/monolog-bundle": "~2.4",
-        "sensio/distribution-bundle": "^3.0.36|^4.0.6|^5.0",
-        "sensio/generator-bundle": "^2.3|^3.0",
-        "incenteev/composer-parameter-handler": "~2.0",
-        "tedivm/stash-bundle": "~0.4",
-        "ezsystems/ezpublish-kernel": "~6.7@dev",
-        "ezsystems/repository-forms": "^1.4",
-        "ezsystems/ezplatform-solr-search-engine": "^1.1@dev",
-        "ezsystems/platform-ui-bundle": "~1.5@dev",
-        "ezsystems/ez-support-tools": "~0.1.0",
+        "symfony/symfony": "^3.2.2",
+        "twig/extensions": "^1.4.1",
+        "symfony/assetic-bundle": "^2.8.1",
+        "symfony/swiftmailer-bundle": "^2.4.2",
+        "symfony/monolog-bundle": "^3.0.3",
+        "sensio/distribution-bundle": "^5.0.18",
+        "incenteev/composer-parameter-handler": "^2.1.2",
+        "ezsystems/ezpublish-kernel": "dev-sf_cache as 6.99.x-dev",
+        "ezsystems/repository-forms": "^1.5",
+        "ezsystems/ezplatform-solr-search-engine": "^1.3@dev",
+        "ezsystems/platform-ui-bundle": "~1.7@dev",
+        "ezsystems/ez-support-tools": "~0.1.2",
         "ezplatform-i18n/ezplatform-i18n-ach_ug": "^1.0@dev",
-        "egulias/listeners-debug-command-bundle": "~1.9",
-        "white-october/pagerfanta-bundle": "1.0.*",
-        "hautelook/templated-uri-bundle": "~1.0 | ~2.0",
-        "doctrine/doctrine-bundle": "~1.3",
-        "symfony/expression-language": "~2.4",
-        "sensio/framework-extra-bundle": "~3.0",
+        "egulias/listeners-debug-command-bundle": "^1.9.1",
+        "white-october/pagerfanta-bundle": "^1.0.7",
+        "hautelook/templated-uri-bundle": "^2.0.1",
+        "doctrine/doctrine-bundle": "^1.6.7",
+        "sensio/framework-extra-bundle": "^3.0.2",
         "ezsystems/content-on-the-fly-prototype": "~0.1.7",
         "willdurand/js-translation-bundle": "^2.7@dev"
     },
     "require-dev": {
         "ezsystems/ezplatform-i18n": "^1.0@dev",
-        "behat/behat": "~3.0",
-        "behat/symfony2-extension": "~2.0",
+        "behat/behat": "^3.3.0",
+        "behat/symfony2-extension": "^2.1.1",
         "behat/mink-extension": "*",
         "behat/mink-goutte-driver": "*",
         "behat/mink-selenium2-driver": "*",
-        "jarnaiz/behat-junit-formatter": "^1.3",
-        "ezsystems/behatbundle": "^6.3"
+        "sensio/generator-bundle": "^3.0",
+        "symfony/phpunit-bridge": "^3.0"
     },
     "suggest": {
         "ezsystems/legacy-bridge": "Provides the full legacy backoffice and legacy features"
@@ -75,12 +76,14 @@
         ]
     },
     "config": {
-        "process-timeout": 3000,
-        "bin-dir": "bin"
+        "process-timeout": 3000
     },
     "extra": {
         "symfony-app-dir": "app",
+        "symfony-bin-dir": "bin",
+        "symfony-var-dir": "var",
         "symfony-web-dir": "web",
+        "symfony-tests-dir": "tests",
         "___symfony-assets-install": "One of 'symlink', 'relative' (symlinks) or 'hard'",
         "symfony-assets-install": "relative",
         "incenteev-parameters": {
@@ -88,7 +91,7 @@
             "file": "app/config/parameters.yml"
         },
         "branch-alias": {
-            "dev-master": "1.8.x-dev"
+            "dev-master": "2.0.x-dev"
         }
     }
 }
