@@ -15,6 +15,6 @@ use Composer\Autoload\ClassLoader;
  */
 $loader = require __DIR__ . '/../vendor/autoload.php';
 
-AnnotationRegistry::registerLoader(array($loader, 'loadClass'));
+AnnotationRegistry::registerLoader([$loader, 'loadClass']);
 
 return $loader;
