@@ -15,17 +15,11 @@ if (($useDebugging = getenv('SYMFONY_DEBUG')) === false || $useDebugging === '')
     $useDebugging = $environment === 'dev';
 }
 
-// Depending on SYMFONY_CLASSLOADER_FILE use custom class loader, otherwise use bootstrap cache, or autoload in debug
-if ($loaderFile = getenv('SYMFONY_CLASSLOADER_FILE')) {
-    require_once $loaderFile;
-} else {
-    require_once __DIR__ . '/../app/autoload.php';
-    if (!$useDebugging) {
-        require_once __DIR__ . '/../app/bootstrap.php.cache';
-    }
+/** @var \Composer\Autoload\ClassLoader $loader */
+$loader = require __DIR__.'/../app/autoload.php';
+if (!$useDebugging) {
+    include_once __DIR__ . '/../var/bootstrap.php.cache';
 }
-
-require_once __DIR__ . '/../app/AppKernel.php';
 
 if ($useDebugging) {
     Debug::enable();
